@@ -34,6 +34,7 @@ from typing import Any
 from typing import Callable
 from typing import Dict
 from typing import Generic
+from typing import Literal
 from typing import Optional
 from typing import overload
 from typing import Protocol
@@ -42,7 +43,6 @@ from typing import TypeVar
 from typing import Union
 
 from . import exc
-from .util.typing import Literal
 
 _T = TypeVar("_T", bound=Any)
 _TCov = TypeVar("_TCov", bound=Any, covariant=True)
