@@ -15,6 +15,7 @@ import typing
 from typing import Any
 from typing import cast
 from typing import List
+from typing import Literal
 from typing import Optional
 from typing import Set
 from typing import Type
@@ -34,7 +35,6 @@ from .. import exc
 from .. import util
 from ..util import chop_traceback
 from ..util import queue as sqla_queue
-from ..util.typing import Literal
 
 if typing.TYPE_CHECKING:
     from ..engine.interfaces import DBAPIConnection
