@@ -8,12 +8,12 @@ from __future__ import annotations
 
 from typing import Any
 from typing import Generic
+from typing import Literal
 from typing import Optional
 from typing import TYPE_CHECKING
 from typing import TypeVar
 
 from .. import util
-from ..util.typing import Literal
 
 if TYPE_CHECKING:
     from ._typing import _PropagateAttrsType
