@@ -3912,7 +3912,7 @@ class DeclOrmForms(fixtures.TestBase):
         {},
         {"repr": False},
         {"repr": True},
-        ({"kw_only": True}, testing.requires.python310),
+        {"kw_only": True},
         {"init": False},
         {"default_factory": True},
         argnames="field_kw",
