@@ -1164,10 +1164,8 @@ class EmbeddedSubqTest(
                 testing.skip_test("python platform not available")
             elif util.py311:
                 int_within_variance(39996, total_size(ck), 0.05)
-            elif util.py310:
-                int_within_variance(29796, total_size(ck), 0.05)
             else:
-                testing.skip_test("python platform not available")
+                int_within_variance(29796, total_size(ck), 0.05)
 
 
 class WithExpresionLoaderOptTest(DeclarativeMappedTest):
